career_agent/pyproject.toml
[project]
name = "career-agent-py"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chainlit>=2.6.0",
    "google-generativeai>=0.8.5",
    "openai-agents>=0.1.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "streamlit>=1.46.1",
    "typing-extensions>=4.14.1",
]
